backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "uvicorn>=0.30.1",
    "playwright>=1.44.0",
    "Pillow>=10.3.0",
    "google-generativeai>=0.5.0",
    "python-dotenv>=1.0.0", 
    "openai>=1.0.0", 
    "anthropic>=0.28.0"
]
